requirements.txt
discord.py>=2.3.0
pytz>=2023.3
aiohttp>=3.8.0
